mavlink: move DISTANCE_SENSOR to streams header
